fixed datatype of rating from int to decimal. implemented businessItemAdapter.

+++ EatMe/app/src/main/java/com/example/bravodavid56/eatme/data/BusinessItem.java
@@ -15,11 +15,11 @@ public class BusinessItem {
     private String display_phone;
     private int review_count;
     private String address; // the data type from the json is initially a list, but we can extract the address string
-    private int rating;
+    private double rating;
     private String categories; // need to create a method to convert the list to string and string to list
 
 
-    public BusinessItem(String id, String name, String image_url, String url, String display_phone, int review_count, String address, int rating, String categories) {
+    public BusinessItem(String id, String name, String image_url, String url, String display_phone, int review_count, String address, double rating, String categories) {
         this.id = id;
         this.name = name;
         this.image_url = image_url;
@@ -100,7 +100,7 @@ public class BusinessItem {
         this.address = address;
     }
 
-    public int getRating() {
+    public double getRating() {
         return rating;
     }
